Split `Sync15Service` into `StorageClient` and `Collections`.
* `StorageClient` seems like a good API boundary. It's a fairly basic
  HTTP client that makes authenticated requests to and decodes
  responses from the storage node.
* I'm less sure about `Collections`. Some of its functionality
  could be handled by the Sync state machine; the uploading logic
  could be moved into `StorageClient`.
